requirements.txt
transformers==4.49.0
trl
datasets
torch
pyyaml
peft
bitsandbytes
accelerate
einops
sentencepiece
